ensure_homebrew_prefix_usable: Provide context for password prompt
If we're going to prompt for a password, we ought to tell people
what we plan to do. If they trust us, they can enter their password,
otherwise, they can bail and then do it themselves.

Also provide some "here's what we're doing, here's what we did"
signposting. It's not THE UNIX(TM) WAY, but some of these functions are
REALLY slow to execute. (Switching to a Makefile, which autologs the
commands that it's going to run unless suppressed with `@`,
might be more NIXy.)